requirements.txt
PyQt5==5.14.1
PyQt5-sip==12.7.1
PyQtChart==5.14.0
pyserial==3.4
